440-issue keep a record of the gas last observed tx spend , and  use that for next transaction we send out.

+++ bifrost/pkg/chainclients/bitcoin/transaction_fee.go
@@ -3,5 +3,5 @@ package bitcoin
 // TransactionFee on bitcoin
 type TransactionFee struct {
 	Fee   float64 `json:"fee"`
-	VSize int64   `json:"v_size"`
+	VSize int32   `json:"v_size"`
 }
