+++ bifrost/pkg/chainclients/bitcoin/utxo_accessor.go
@@ -5,6 +5,6 @@ type UnspentTransactionOutputAccessor interface {
 	GetUTXOs() ([]UnspentTransactionOutput, error)
 	AddUTXO(UnspentTransactionOutput) error
 	RemoveUTXO(key string) error
-	UpsertTransactionFee(fee float64, vSize int64) error
-	GetTransactionFee() (float64, int64, error)
+	UpsertTransactionFee(fee float64, vSize int32) error
+	GetTransactionFee() (float64, int32, error)
 }
